Move attachment to protocols.compliance
