--- protocols/attachment/doc.go
@@ -1,3 +1,0 @@
-// Package attachment implements Stellar Attachment Convention
-// https://www.stellar.org/developers/guides/attachment.html
-package attachment
