+++ protocols/compliance/attachment.go
@@ -1,4 +1,4 @@
-package attachment
+package compliance
 
 import (
 	"crypto/sha256"
