+++ protocols/compliance/sender_info_test.go
@@ -1,4 +1,4 @@
-package attachment
+package compliance
 
 import (
 	"testing"
